refactor(testes): renomeia pacote de testes de entity_test para entity
