+++ internal/domain/entity/gyroscope.go
