+++ internal/domain/entity/gyroscope_test.go
@@ -1,14 +1,13 @@
-package entity_test
+package entity
 
 import (
-	entity "github.com/HaroldoFV/desafio/internal/domain"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
 func TestNewGyroscope(t *testing.T) {
 	t.Run("Valid Gyroscope", func(t *testing.T) {
-		gyro, err := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
+		gyro, err := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
 		require.Nil(t, err)
 		require.NotNil(t, gyro)
 		require.Equal(t, "MPU-6050", gyro.GetName())
@@ -17,36 +16,36 @@ func TestNewGyroscope(t *testing.T) {
 	})
 
 	t.Run("Invalid Name", func(t *testing.T) {
-		_, err := entity.NewGyroscope("", "InvenSense GY-521", "00:11:22:33:44:55")
+		_, err := NewGyroscope("", "InvenSense GY-521", "00:11:22:33:44:55")
 		require.EqualError(t, err, "name cannot be empty")
 	})
 
 	t.Run("Name Too Long", func(t *testing.T) {
 		longName := string(make([]byte, 101))
-		_, err := entity.NewGyroscope(longName, "InvenSense GY-521", "00:11:22:33:44:55")
+		_, err := NewGyroscope(longName, "InvenSense GY-521", "00:11:22:33:44:55")
 		require.EqualError(t, err, "name cannot be longer than 100 characters")
 	})
 
 	t.Run("Invalid Model", func(t *testing.T) {
-		_, err := entity.NewGyroscope("MPU-6050", "", "00:11:22:33:44:55")
+		_, err := NewGyroscope("MPU-6050", "", "00:11:22:33:44:55")
 		require.EqualError(t, err, "model cannot be empty")
 	})
 
 	t.Run("Model Too Long", func(t *testing.T) {
 		longModel := string(make([]byte, 51))
-		_, err := entity.NewGyroscope("MPU-6050", longModel, "00:11:22:33:44:55")
+		_, err := NewGyroscope("MPU-6050", longModel, "00:11:22:33:44:55")
 		require.EqualError(t, err, "model cannot be longer than 50 characters")
 	})
 
 	t.Run("Invalid MAC Address", func(t *testing.T) {
-		_, err := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "invalid-mac")
+		_, err := NewGyroscope("MPU-6050", "InvenSense GY-521", "invalid-mac")
 		require.EqualError(t, err, "invalid MAC address format")
 	})
 
 }
 
 func TestGyroscope_GetMethods(t *testing.T) {
-	gyro, _ := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:1A:2B:3C:4D:5E")
+	gyro, _ := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:1A:2B:3C:4D:5E")
 
 	t.Run("GetID", func(t *testing.T) {
 		require.NotEmpty(t, gyro.GetID())
@@ -83,7 +82,7 @@ func TestGyroscope_GetMethods(t *testing.T) {
 
 func TestGyroscope_Update(t *testing.T) {
 	t.Run("Update with valid data", func(t *testing.T) {
-		gyro, _ := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
+		gyro, _ := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
 		err := gyro.Update("L3GD20H", "STMicroelectronics", "00:AA:BB:CC:DD:EE")
 		require.Nil(t, err)
 		require.Equal(t, "L3GD20H", gyro.GetName())
@@ -92,20 +91,20 @@ func TestGyroscope_Update(t *testing.T) {
 	})
 
 	t.Run("Update with empty name", func(t *testing.T) {
-		gyro, _ := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
+		gyro, _ := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
 		err := gyro.Update("", "STMicroelectronics", "00:AA:BB:CC:DD:EE")
 		require.EqualError(t, err, "name cannot be empty")
 	})
 
 	t.Run("Update with too long name", func(t *testing.T) {
-		gyro, _ := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
+		gyro, _ := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
 		longName := string(make([]byte, 101))
 		err := gyro.Update(longName, "STMicroelectronics", "00:AA:BB:CC:DD:EE")
 		require.EqualError(t, err, "name cannot be longer than 100 characters")
 	})
 
 	t.Run("Update with invalid MAC address", func(t *testing.T) {
-		gyro, _ := entity.NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
+		gyro, _ := NewGyroscope("MPU-6050", "InvenSense GY-521", "00:11:22:33:44:55")
 		err := gyro.Update("L3GD20H", "STMicroelectronics", "invalid-mac")
 		require.EqualError(t, err, "invalid MAC address format")
 	})
